server-cropcare/requirements.txt
tensorflow~=2.18.0
numpy~=2.0.2
Flask~=3.1.0
pillow~=11.1.0
requests~=2.28.1
flask-cors~=3.1.1
